// client-monorepo/libs/applets/insurance/src/lib/app/features/vehicle/features/third-party/data-access/enums/third-party-routes.enum.ts
export enum ThirdPartyRoutesEnum {
  ThirdParty = 'third-party',
  CarInfo = 'car-info',
  Sanhab = 'sanhab',
  Usage = 'usage',
  ExInsurer = 'ex-insurer',
  Date = 'date',
  ExInsurerInfo = 'ex-insurer-info',
  Discount = 'discount',
  Detail = 'detail',
  Claim = 'claim',
  PlateList = 'plate-list',
  PriceCardList = 'price-card-list',
  Order = 'order',
  Checkout = 'checkout',
  PrepareDetail = 'prepare-detail',
  Error = 'error',
  Payment = 'payment',
  UserInfo = 'user-info',
  Address = 'address',
  UploadDocument = 'upload-document',
  AdditionalUploadDocument = 'additional-upload-document',
  GetPolicyMethod = 'get-policy-method',
  Complete = 'complete',
  State = 'state',
  Result = 'result',
  GoToPayment = 'go-to-payment',
  ResolveDocumentsConflict = 'resolve-documents-conflict',
  PLPCardSelect = 'plp-card-select',
  VerifyPostalCode = 'verify-postal-code'
}