// client-monorepo/libs/applets/insurance/src/lib/app/data-access/enums/third-party-page-titles.enum.ts
export enum ThirdPartyPageTitlesEnum {
  Plate = 'Plate',
  Sanhab = 'Sanhab',
  SanhabCarInfo = 'SanhabCarInfo',
  CarInfo = 'CarInfo',
  CarInfoClaim = 'CarInfoClaim',
  CarInfoClaimDetail = 'CarInfoClaimDetail',
  CarInfoDiscount = 'CarInfoDiscount',
  CarInfoDiscountDetail = 'CarInfoDiscountDetail',
  CarInfoExInsurer = 'CarInfoExInsurer',
  CarInfoExInsurerDate = 'CarInfoExInsurerDate',
  CarInfoExInsurerInfo = 'CarInfoExInsurerInfo',
  PLP = 'PLP',
  PLPCardSelect = 'PLPCardSelect',
  OrderCheckout = 'OrderCheckout',
  OrderCompleteOrder = 'OrderCompleteOrder',
  OrderGetPolicyMethod = 'OrderGetPolicyMethod',
  OrderState = 'OrderState',
  OrderResolveDocumentsConflict = 'OrderResolveDocumentsConflict',
  OrderUploadDocument = 'OrderUploadDocument',
  OrderAdditionalUploadDocument = 'OrderAdditionalUploadDocument',
  OrderUserAddress = 'OrderUserAddress',
  OrderVerifyPostalCode = 'OrderVerifyPostalCode',
  OrderUserInfo = 'OrderUserInfo',
  PaymentResult = 'PaymentResult',
  PaymentGoToPayment = 'PaymentGoToPayment',
  PaymentCheckHybrid = 'PaymentCheckHybrid',
  login = 'Login',
}

export const THIRD_PARTY_PAGE_TILES_TRANSLATOR = {
  [ThirdPartyPageTitlesEnum.Plate]: 'پلاک',
  [ThirdPartyPageTitlesEnum.Sanhab]: 'اطلاعات خودرو',
  [ThirdPartyPageTitlesEnum.SanhabCarInfo]: 'مشخصات دریافت شده',
  [ThirdPartyPageTitlesEnum.CarInfo]: 'مشخصات دریافت شده',
  [ThirdPartyPageTitlesEnum.CarInfoClaim]: 'خسارت',
  [ThirdPartyPageTitlesEnum.CarInfoClaimDetail]: 'جزییات خسارت',
  [ThirdPartyPageTitlesEnum.CarInfoDiscount]: 'تخفیف',
  [ThirdPartyPageTitlesEnum.CarInfoDiscountDetail]: 'جزییات تخفیف',
  [ThirdPartyPageTitlesEnum.CarInfoExInsurer]: 'بیمه‌گر قبلی',
  [ThirdPartyPageTitlesEnum.CarInfoExInsurerDate]: 'تاریخ بیمه‌گر قبلی',
  [ThirdPartyPageTitlesEnum.CarInfoExInsurerInfo]: 'مشخصات بیمه‌گر قبلی',
  [ThirdPartyPageTitlesEnum.PLP]: 'انتخاب بیمه‌گر',
  [ThirdPartyPageTitlesEnum.PLPCardSelect]: 'در حال انتخاب بیمه‌گر',
  [ThirdPartyPageTitlesEnum.OrderCheckout]: 'جزییات سفارش',
  [ThirdPartyPageTitlesEnum.OrderCompleteOrder]: 'تکمیل سفارش',
  [ThirdPartyPageTitlesEnum.OrderGetPolicyMethod]: 'نحوه دریافت بیمه‌نامه',
  [ThirdPartyPageTitlesEnum.OrderState]: 'وضعیت سفارش',
  [ThirdPartyPageTitlesEnum.OrderResolveDocumentsConflict]: 'آپلود مدارک',
  [ThirdPartyPageTitlesEnum.OrderUploadDocument]: 'آپلود مدارک',
  [ThirdPartyPageTitlesEnum.OrderAdditionalUploadDocument]: 'آپلود مدارک',
  [ThirdPartyPageTitlesEnum.OrderUserAddress]: 'آدرس',
  [ThirdPartyPageTitlesEnum.OrderVerifyPostalCode]: 'ثبت کد پستی',
  [ThirdPartyPageTitlesEnum.OrderUserInfo]: 'مشخصات کاربر',
  [ThirdPartyPageTitlesEnum.PaymentResult]: 'نتیجه پرداخت',
  [ThirdPartyPageTitlesEnum.PaymentGoToPayment]: 'نتیجه پرداخت',
  [ThirdPartyPageTitlesEnum.PaymentCheckHybrid]: 'نتیجه پرداخت',
  [ThirdPartyPageTitlesEnum.login]: 'لاگین',
};